\name{JR1989}
\alias{JR1989}
\docType{data}
\title{
Simulated Johnson and Rusbult (1989) Data on the Effects of Relationship Commitment and Target Attractiveness on Ratings of Attraction to Target
}
\description{
A simulated data set based on the means reported by Johnson and Rusbult (1989).
}
\usage{data(JR1989)}
\format{
  A data frame with 200 observations on the following 4 variables.
  \describe{
    \item{\code{subj}}{Subject identification number}
    \item{\code{Commitment}}{Subject's level of commitment to their current romantic relationship}
    \item{\code{Attractiveness}}{Attractiveness of target person}
    \item{\code{rating}}{Average of 3 ratings: (1) to what degree are you attracted to X (1 = not at all, 9 = extremely), (1) how do you think a relationship with X would compare to your current relationship (1 = much worse, 9 = much better), and (1) to what extent do you think you would have a satisfying relationship with X (1 = not at all satisfying, 9 = extremely satisfying) }
  }
}
\details{
These results are sometimes used as a demonstration of the tendency to derogate unselected alternatives when commitment is high.  Subjects reported their own level of commitment in a romantic relationship and then gave 3 ratings for a photograph that had been pre-tested to be either moderately attractive or very attractive.
}
\source{
Johnson, D. & Rusbult, C. (1989). Resisting temptation: Devaluation of alternative partners as a means of maintaining commitment in close relationships. Journal of Personality and Social Psychology, 57, 967-980.
}
