\name{multitask}
\alias{multitask}
\docType{data}
\title{
Gender differences in multi-tasking ability
}
\description{
Data from a student experiment on gender differences in multi-tasking ability
}
\usage{data(multitask)}
\format{
  A data frame with 22 observations on the following 11 variables.
  \describe{
    \item{\code{Participant}}{a factor of unique subject identification numbers}
    \item{\code{Gender}}{subject gender, a factor with levels \code{male} \code{female}}
    \item{\code{Age}}{subject age in years, a numeric vector}
    \item{\code{Year}}{year in school, a factor with levels \code{freshman} \code{sophmore} \code{junior} \code{senior}}
    \item{\code{Greek}}{member of a Greek fraternity or sorority, a factor with levels \code{yes} \code{no}}
    \item{\code{GPA}}{grade point average on 4-point scale with 4 = perfect A average, 3 = B average, etc., a numeric vector}
    \item{\code{EC}}{did subject complete this study for extra credit? A factor with levels \code{yes} \code{no}}
    \item{\code{MathOne}}{performance on a math test in the single-task condition}
    \item{\code{MathTwo}}{performance on a math test in the multi-task condition}
    \item{\code{MovieOne}}{performance on a test of movie details in the single-task condition}
    \item{\code{MovieTwo}}{performance on a test of movie details in the multi-task condition}
  }
}
\details{
Materials
Three tasks were used to measure participants' ability to multitask: watching a 
video, taking a written math test, and responding to instant messages. The math questions 
consisted of basic algebra and other high school-level questions taken from the ACT 
practice tests (The College Board, 2007). The films involved were Disney films, 
Pocahontas and The Hunchback of Notre Dame, and participants were asked 
questions regarding details that would need to be observed in the clips. A researcher in 
another room used a laptop connected to the internet via campus wireless 
network connection to communicate with participants.  Instant messages, a 
communication tool that allows users to communicate via typed messages transmitted 
instantly to another user, were transmitted via AOL Instant Messenger 
software. Participants had been told to expect a series of messages and to respond as 
quickly as possible through the trial. Messages covered a broad range of topics, 
including:  "What is your favorite color?" "Do you have a favorite sports team?" or "How 
many siblings do you have?"
Participants viewed five-minute clips from Disney's Pocahontas and The 
Hunchback of Notre Dame (-1:08:10 to -1:03:25) shown on a 32 inch RCA television and 
completed a sheet of ten questions related to each clip. Some example questions include 
"What animal danced near the gypsy" and "What was the rank of the gentleman in blue?" 
The films used in the single-task condition and the multi-task condition were alternated 
with each trial.
The math assessment consisted of 40 sample ACT Math questions (The College 
Board, 2007). Problems were selected with the goal of finding a set that participants 
would be able to complete without the aide of a calculator.  Questions were divided into 
two groups of 20: one for the single-task trial and one for the multi-task trial. The set 
given for each trial alternated with each session.  An example question is, "Which of the 
following is equivalent to (x)(x)(x)(x), for all x ?" with choice of answers 4x, x^4, x+4, 
4^x, or 2x^2.
Additionally, the study included an eight-item demographic questionnaire, which 
participants completed at the end of the session.  The questions included gender, age, 
major, year in school, ethnicity, Greek affiliation, and GPA.
Procedure 
Participants were told the study would examine the effects of multi-tasking on 
work quality.  All participants signed an informed consent form prior to 
participating. Trials were run with eight participants in a group.  Each participant 
completed a single-task trial and a multiple-task trial, with half of the participants being 
randomly assigned to begin with the single task and the other half completing the 
multiple task first. In order to complete the single-task trial, participants were placed in 
an adjacent room, which was empty. Participants sat at tables in pairs of two, each facing 
the front of the room. For the multiple-task trial, participants were seated at a block of 
four computers.  The configuration of the four stations was such that each participant had 
another participant approximately three feet to his/her left or right and was also facing 
another participant at a distance of approximately five feet.   
For the single task, each participant completed 20 math problems, viewed a five 
minute film clip and completed a sheet of questions related to the clip, which was 
collected at the end of the clip. The television was positioned at the front of the room, 
within ten feet of each participant. 
For the multiple-task test, participants completed 20 different but similar math 
problems.  However, they simultaneously responded to instant messages sent from a 
researcher in another room and, while watching a five-minute film clip, completed a 
worksheet asking questions about details from the film clip.  As in the single-task 
condition, the television was positioned at the front of the room approximately ten feet 
from the participants. The worksheet was collected at the end of the film clip, and 
another five minutes were allowed before ending the instant message conversation and 
collecting the math problem worksheets.
After participants completed both the single- and multiple-tasks, they filled out a 
short survey sheet with basic demographic information, including gender, year in college, 
GPA, and age.  Participants were given a written debriefing form, thanked for their 
participation, offered cookies, and dismissed. 
}
\source{
Bauer, K., DeVincentis, D., & Taber, J. (2008). Gender differences in the effects of multi-tasking performance. Unpublished manuscript, Hanover College, Hanover, IN.
Available online at:
http://vault.hanover.edu/~altermattw/methods/assets/posterpics/Winter2008/Bauer-Devincentis-Taber.pdf
}
\keyword{datasets}
