\name{nss}
\alias{nss}
\docType{data}
\title{
Neo-Sexism Scale Responses from 189 Students
}
\description{
Responses by 189 students to Tougas et al.'s (1995) Neo-Sexism Scale made using a 5-point Likert scale from 1 (Strongly Disagree) to 5 (Strongly Agree)
}
\usage{data(nss)}
\format{
  A data frame with 189 observations on the following 12 variables.
  \describe{
    \item{\code{ID}}{Unique subject identification number}
    \item{\code{q01}}{Discrimination against women in the labor force is no longer a problem in the U.S.}
    \item{\code{q02}}{I consider the present employment system to be unfair to women.}
    \item{\code{q03}}{Women shouldn't push themselves where they are not wanted.}
    \item{\code{q04}}{Women will make more progress by being patient and not pushing too hard for change.}
    \item{\code{q05}}{It is difficult to work for a female boss.}
    \item{\code{q06}}{Women's requests in terms of equality between the sexes are simply exaggerated.}
    \item{\code{q07}}{Over the past few years, women have gotten more from the government than they deserve.}
    \item{\code{q08}}{Universities are wrong to admit women in costly programs such as medicine, when in fact a large number will leave their jobs after a few years to raise their children.}
    \item{\code{q09}}{In order not to appear sexist, many men are inclined to overcompensate women.}
    \item{\code{q10}}{Due to social pressures, firms frequently have to hire underqualified women.}
    \item{\code{q11}}{In a fair employment system, men and women would be considered equal.}
  }
}
\details{
To understand the contribution made by this scale, it is useful to consider 
Benokraitis and Feagin's (1986) delineation of three forms of sexism:  overt, 
covert, and subtle.  Overt sexism is "unequal and harmful treatment of women 
that is readily apparent, visible, and observable, and can be easily documented" 
(Benokraitis & Feagin, 1986, p. 30).  With its straightforward and relatively 
transparent items, the Attitudes Toward Women Scale likely taps this type of 
sexism (Swim & Cohen, 1997).  Covert sexism involves "engaging in unequal and 
harmful treatment of women and men in a hidden or clandestine manner" (Swim & 
Cohen, 1997, p. 104).  Individuals engaging in covert sexism may say that they 
endorse gender equality, but then act in a way that may "intentionally undermine 
women's work or lead them to fail" (Swim & Cohen, 1997, p. 104).  Whereas overt 
sexism is easily identified by behavior that is harmful to women, covert sexism 
is defined more by the intention to harm women while avoiding detection.  Two 
scales have been designed to assess individuals' covert sexist beliefs:  the 
Modern Sexism scale (Swim, Aikin, Hall, & Hunter, 1995) and the Neosexism scale 
(Tougas et al., 1995).  These scales measure whether respondents tend to (a) 
deny the existence of discrimination against women, (b) resent complaints about 
discrimination, and (c) resent special "favors" for women.  Campbell, Schellenberg,
and Senn (1997) found that the Neosexism Scale had higher internal validity and 
exhibited greater gender differences than the Modern Sexism Scale.  
}
\source{
Altermatt, T. W. (2001). Chivalry: The relation between a cultural script and sex stereotypes. (Unpublished doctoral dissertation). University of Illinois, Champaign-Urbana.
}
\references{
Benokraitis, N., & Feagin, J. (1986). Modern sexism. Englewood Cliffs, NJ: Prentice-Hall.
Campbell, B., Schellenberg, E. G., & Senn, C. (1997). Evaluating measures of contemporary sexism. Psychology of Women Quarterly, 21, 89-102.
Swim, J. K., Aikin, K. J., Hall, W. S., & Hunter, B. A. (1995). Sexism and racism: Old-fashioned and modern prejudices. Journal of Personality and Social Psychology, 68, 199-214.
Swim, J., & Cohen, L. (1997). Overt, covert, and subtle sexism: A comparison between the Attitudes Toward Women and Modern Sexism Scales. Psychology of Women Quarterly, 21, 103-118.
Tougas, F., Brown, R., Beaton, A. M., & Joly, S. (1995). Neosexism: Plus ca change, plus c'est pareil. Personality and Social Psychology Bulletin, 21, 842-849.
}

