\name{Strohmetz1}
\alias{Strohmetz1}
\docType{data}
\title{
Simulated data from Strohmetz et al. (2002) Experiment 1
}
\description{
Simulated data using the means and standard deviations reported from Strohmetz et al. (2002) in Experiment 1.
}
\usage{data(Strohmetz1)}
\format{
  A data frame with 92 observations on the following 2 variables.
  \describe{
    \item{\code{Condition}}{a character vector indicating whether the dining party received chocolate 
with their bill or whether they received no chocolate.}
    \item{\code{Tip_Percentage}}{a numeric vector equal to the gratuity divided by the pre-tax bill, 
multiplied by 100. So, a 17 corresponds to a 17 percent tip.}
  }
}
\details{
From the article: Participants. Participants were 92 dining parties (individual party sizes were not recorded) at a 
restaurant in downtown Ithaca, New York. The restaurant seats approximately 66 people and provided
a full dinner menu, including alcohol and a casual dinner atmosphere.
Procedure. The experiment was conducted over two weekday and two weekend dinner shifts during the
fall of 1992. Two seasoned male waiters served as experimental accomplices. Just prior to delivering
the check to a table, these waiters selected a playing card from a shuffled deck of cards. If the
card was red, they gave each person in the dining party a fancy, foil-wrapped piece of chocolate
when they delivered the check. This fancy piece of chocolate was used to differentiate this token
gift from the inexpensive wrapped mints that are often provided by restaurants. When the card was 
black, the servers delivered the check without giving the guests pieces of chocolate. Upon delivery
of the check, the server thanked the customers and recorded the total guest check, the amount of the
gratuity, the experimental condition, the method of payment (cash vs. credit), and the gender of the
bill payer. At the end of the experiment, two or three dining parties were assigned to one condition
in order to achieve equal sample sizes for the experimental and control conditions.
Tip percentage was determined by dividing the amount of the tip by the size of the bill before 
taxes and multiplied by 100. It was hypothesized that the servers would receive larger tip percentages
when they gave each member of the dining party a piece of chocolate along with the final check. 
}
\source{
Strohmetz, D. B., Rind, B., Fisher, R., & Lynn, M. (2002). Sweetening the till: The use of candy
to increase restaurant tipping. Journal of Applied Social Psychology, 32, 2, 300-309.
}

